chore(superchain): stop pre-populating package cache (#732)
The install of dotnet in the Docker build caused the first-time
experience to trigger as the environment variable apparently must be set
to `true` and not `1`. This causes a package cache to be pre-populated
(to spee dup future builds), however as this is baked into a Docker
image, this will cause bloat & will probably result in stale cached
artifacts.