fix(kernel): correctly serialize enum values
When string-valued enums were serialized, an incorrect reference would
be created if the label of the enum entry did not match it's string
value; due to the lookup code being incorrect.

Updated a string-valued enum and introduced new compliance tests to
ensure such enum values are passed across correctly.

Fixes #753