fix(python): clear error message when trying to serialize function
A common error seems to be to try to pass a function as a parameter
into a JSII call (have seen 3 occurrences of this in the last week).

Previously, this call would fail when trying to assing the
`__jsii_type__` property to the function object.

Specifically catch this case to give a helpful error message.

Fixes aws/aws-cdk#4064.