chore: migrate mono-repository to use "yarn workspaces"
This appears to improve the bootsrapping performance and makes the lock
files much easier to manager (`yarn` maintains a single `yarn.lock` file
at the root of the mono-repository instead of having one for each
package like `npm` does).

As required, this adjusted the `jsii/superchain` Docker image to also
contain `yarn`, and changes how the `node` binary is installed to using
the `nodesource` repository (so as to satisfy the `yarn` RPM
dependencies without hacking around too much).