chore: migrate mono-repository to use "yarn workspaces" (#891)
* chore: migrate mono-repository to use "yarn workspaces"

This appears to improve the bootsrapping performance and makes the lock
files much easier to manager (`yarn` maintains a single `yarn.lock` file
at the root of the mono-repository instead of having one for each
package like `npm` does).

As required, this adjusted the `jsii/superchain` Docker image to also
contain `yarn`, and changes how the `node` binary is installed to using
the `nodesource` repository (so as to satisfy the `yarn` RPM
dependencies without hacking around too much).

* CodeBuild: install yarn if it is not there already (temporarily)

* Temporarily skip yarn's engine check

* better way to tune the ignore-engines

* fixup Gemfile.lock

* add missing yarn install

* added required link for bundled dependency

* fix hoisting problem w/ jsii-calc-bundled, use * versions where possible.

* increase timeout of negatives test (need to investigate why they are so slow though)

* fix build-test

* fixup python test

* try to install & test from the same container (suspecting ruby's .bundle directory somehow gets lost)

* unset some problematic BUNDLER_ environment variables in Docker (they assumed single-app workloads, we're making a development image)

* revert * dependency versions