feat(java): offer Builders for certain Java classes
Provide a nested `Builder` class with a static `create` factory for all
classes that adhere to the following constraints:
- The class is not `abstract`
- The class has a visible `initializer` (aka `constructor`)
- The constructor has exactly one `struct` (aka `datatype`) parameter
- The constructor is not variadic

Such builders will request the *positional* parameters to be passed
directly to the `Builder#create` method and can (at least currently) not
be subsequently modified. All parameters from the `struct` are then set
direcly on the `Builder` instance.

Fixes #488