chore: upgrade to @types/node 10.x (#894)
Actually upgrade the `node` engine support to 10.x tree, as has
been declared in the `package.json` files for a month now.

Additionally, standardized the `compilerOptions` in the `tsconfig.json` files
that are *not* generated by `jsii`, in order to use `ES2018` standard library,
enable `strict` mode (this includes `strictPropertyInitialization`, which was not
enabled previously and required a few code changes in `jsii-pacmak` that were
made ad-minima).

The library clause of  `tsconfig.json` files that are generated by `jsii` has not
been changed yet. This means you need to use `node >= 10.x` in order to use
`jsii`, but you cannot use it to build code that depend on APIs that are not yet
available in `node 8`. This behavior will be updated in a second phase.

Related: #794