fix(java,dotnet): emit default implementations for optional properties
For optional properties on interfaces, emit default implementations when
possible (for .NET, this requires `netcoreapp3.0` or later) so that
adding a new optional property on an interface does not break existing
core that implements the interface.

In it's current form, the getters always return `null`, while the
setters throw an exception (this may change in the future).

Fixes #543