fix(pacmak): occasional EISDIR failure
There was a race condition in capturing the output of `npm pack`.
Ocassionally we would miss the output, not get a tarball name, combine
the empty string with a directory, and then get an EISDIR error when
trying to copy the tarball as a file (but giving it a directory name).

This race has been in there forever, but it became easier to trigger
since we started running a lot of `npm pack`s in parallel.

The solution is to wait for a different event on the `ChildProcess`
object.