fix(runtime): make kernel 'load' operation synchronous (#951)
The `jsii-runtime` process cannot perform async operations while waiting
for a (synchronous) callback to return (the response flow is otherwise
ambiguous). In order to allow loading of dependencies within a calllback
flow, the `load` kernel API needed to be made syncrhonous.

The reason why this is needed is because the host languages runtime
libraries may be required to load dependencies at the last minute, for
it may not be possible to determine what dependencies are required
before  a callback is executed.

Introduced a (unit) test to confirm the `jsii-runtime` is indeed able to
perform `load` operations within a callback context.