feat(rosetta): extract and compile samples into "tablets" (#925)
Version 2 of the "sampiler" is called "Rosetta". It now provides
more control over the compilation of samples found in the source code.

It integrates into an existing build by running `jsii-rosetta extract`,
which will extract sample code from a jsii assembly, compile it, convert
it to all supported languages, and storing the result in a a "tablet
file" (effectively, a sample dictionary).

Tablet files can then be used by `jsii-pacmak` to look up translations
for the code samples it encounters as it is generating language-specific
sources.

In case the build does not contain a Rosetta step, Pacmak will try to
convert samples that are not found in the tablet on the fly. However,
the samples will not benefit from compilation, type checking and fixture
support.

This change also fixes the `jsii-pacmak` all-at-once builder, and will
properly copy out binaries for the .NET and Java builds back to the
declared output directories, and properly use output directories of
packages not included in the megabuild as local package repositories.