Refactor tests a bit
Use describe() for files and test() for languages. Use beforeAll() to
set up the translator in a way that Jest can account for the time spent
there, and be sure to clean up afterwards because the thing is huge.

Translate all existing Python tests into the new format, and make sure
TSC won't compile the example TypeScript files.