fix(java): handle null-able collections correctly (#986)
When a `null` value was returned fro an optional collection (List or
Map), it would be passed to the corresponding unmodifiable wrapper
(`java.util.Collections.unmodifiable~`), however these method are not
`null`-safe.

This makes the wrapping conditional to the collection not being `null`.

This addresses the cause of aws/aws-cdk#4316