fix(jsii): unable to depend on modules with private declarations (#995)
Due to a TypeScript [bug], it was impossible to write a jsii module
that takes a dependency on a module that has private type declarations
in their d.ts file since they are untyped by definition.

This temporary fix sets `noImplicitAny` to `false` in order to disable
this check and enable the dependency use case until we upgrade to the 3.7.x
version-line.

Fixes #994

[bug]: https://github.com/microsoft/TypeScript/issues/33894