feat(jsii-config): introducing jsii-config (#981)
Defines jsii-config CLI application for configuring a jsii package's jsii options inside package.json. These options dictate what target languages and language specific packaging options jsii uses when building.

The application uses inquirer for interactive prompts and yargs for argument parsing.

jsii-config will help configure packages with or without an existing jsii configuration in package.json. If the object does exist, the default values of fields that are prompted will be the existing values. This allows users to just hit 'enter' until they get to the options they wish to change and modify only those.

After the user inputs all of the values for the relevant prompts, they are asked to confirm the config. If the user selects 'yes', the config is written to package.json, if they select 'no', they will be prompted to fill out the fields again, however, the default values will now be what they previously entered. This is done so the user will only need to change the specific fields that need it.

Fixes #904