chore: refactor some packages into @jsii namespace
Moving all packages that do not hold an "entry-point" CLI to be hosted
within the `@jsii` npm namespace (this includes private packages that
are responsible for the language-specific runtime libraries), in order
to make the repository a little easier to navigate.

Additionally, ensured the directory structure of the monorepo
(particularly under `packages/`) matches the `npm` package names, here
also to make the repository easier to navigate. This involved moving the
`@scope/` test packages under an `@scope/` directory.

The `jsii-build-tools` package was moved to a `tools/` top-level
directory, matching what is done in the aws/aws-cdk repository. It may
later be refactored into more granular utilities, etc... Additional
utilities will be added in there, such as what is called for by #1075.