fix(java): missing remarks in builder documentation (#1111)
The Javadoc entries for setting methods on the generated `Builder`
classes did not represent the `remarks` section of the `docs` object,
resulting in only partial documentation being available on the builders.

This makes sure the `remarks` section gets rendered correctly in all the
locations that previously ignored it, and the fix is demonstrated good
by virtue of the changes it induced in the regression test improvements.

Fixes #1094