fix(python): correctly emit sligified positional args (#1081)
In the implementation of kernel calls, the positional argument list did
not use the slugified argument name (in case where such an argument had
a name conflict with a lifted kwarg), resulting in an incorrect argument
list to be passed to `jsii-kernel`. This typically resulted in a type
check error in the `jsii-kernel`, but could also have silently succeeded
at making an incorrect call.

Fixes aws/aws-cdk#4302