chore: improve metadata of NuGet packages (#1112)
Convert the `package.json` `keywords` into `PackageTags` in the NuGet
metadata. This is applied on the packages hosted in this repository as
well as generated code.

The `keywords` section from `package.json` is now forwarded into the
`.jsii` assembly to support this feature.

At this point, could not find a way to carry these keywords/tags to
Maven Central or PyPI, so only NuGet actually leverages those.