fix(reflect): failure to load assembly from "npm pack" (#1117)
When the `package.json` file is the result of extracting a tarball
obtained from `npm pack`, it may have been processed by `npmjs.com`
during publishing, which can cause the `bundleDependencies` property to
appear with the value `false`. The null-coalescing (`??`) operator does
not rescue this properly and this results in a `TypeError`.

This is fixed by using `||` chaining instead of a null-coalescing, which
rescues from all falsey values in this case.