fix(dotnet): documentation strings sometimes invalid
The XML-doc generation for the C# classes was not correctly accounting
for the necesaity to encode certain characters (e.g: `&`) for use in XML
context, which resulted in invalid documentation strings being generated.
Additionally, slugified parameter names need to be represented in their
"pure" form (without the `@` slug) in the `<param>` tag.

This uses `xmlbuilder` to generate the minimal XML tag pairs, ensuring
the content is correctly encoded. The `@` slug is rmeoved from `<param>`
tags, and the `<summary>` block is generated inconditionally (even if
empty), silencing a C# warning about there being no documentation XML on
a public member.

Finally, moved some `await` in `for` loops out, to hopefully positively
impact the generation speeds overall.