fix(java,dotnet): abstract properties have concrete implementations (#1128)
* fix(java,dotnet): abstract properties have concrete implementations

The generated code for abstract properties in Java and C# included fully
concrete implementations, instead of an abstract declaration. This made
it possible to subclass those types without actually implementing those
members, resulting in invalid code.

This changes the code generation to actually emit the `abstract` keyword
and not generate a full concrete implementation.

Fixes #240
Fixes #1011

* add some test coverage