feat(rosetta): translate examples to Java and C# (#985)
Update `rosetta` to also emit translations for Java and C#. Testing is
now done differently, by dropping source files into the
`test/translations` directory for each target language.

`pacmak` for C# now generates a `NamespaceDoc` class (which is hidden
from the IDE as much as possible) which contains the (`README`) module
documentation. This can be picked up by the doc generator to generate
namespace-level documentation.

Examples can be compiled in a special suite, and the translated examples
will be annotated depending on whether the example source successfully
compiled or not.

Fixes summary extraction for some variants of docstrings.