fix(jsii): `--watch` causes immediate failure (#1150)
* fix(jsii): `--watch` causes immediate failure

A change in the TypeScript compiler API around how the compiler host
reacts when invoked in `--watch` mode caused a failure in `jsii` due to
the method changing behaviors from not returning at all when watching to
instead returning an object that can be used to inspect/influence the
background watch process.

Changed the behavior of the `Compiler` interface so it now acts as an
`EventEmitter` when operating in `--watch` mode, and introduced a test
that confirms that operating with `--watch` results in the correct
behavior of watching for file changes and dynamically re-compiling.

Fixes #1149

* some PR feedback

* fix dummy project

* better encapsulate watch API

* Update packages/jsii/bin/jsii.ts

Co-Authored-By: Elad Ben-Israel <benisrae@amazon.com>

* expose ts API instead of wrapping it

Co-authored-by: Elad Ben-Israel <benisrae@amazon.com>