fix: poor error message on scalar deserialization type mismatch (#1187)
* fix: poor error message on scalar deserialization type mismatch

When deserializing a scalar (string, boolean, number, ...), and receiving a value that is not
a scalar, the error message would not mention waht kind of scalar value was required, making
the message less useful than it could be.

Used the description of the type reference in the error message so that the specific scalar
type is direction mentioned in the error message there. The description of the incorrect
value that was received remains unchanged (typically the JSON representation of the value),
and may be improved in a subsequent change.

* remove unnecessary escape

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>