feat(superchain): upgrade maven to 3.6.3
This involves installing maven from tarball instead of installing it from
the RPM repositories provided in the AmazonLinux 2 base image. This is
somewhat unfortunate, but is (currently) the only way to get a recent build
of maven without depending on third-party repositories which would need
vetting before they can be safely used.