chore(java): get rid of JSR-305 references (#1263)
Those have been replaced with JetBrains annotations, which does not have the
possible licensing issues that JSR-305 has, and is generally speaking still being
maintained (while JSR-305 has been dormant for years now).

While touching on the `pom.xml` files, upgraded a couple of dependencies
that had newer and better versions, added the necessary plug-ins to make
sure this is easy to maintain going forward (`versions-maven-plugin`), and
that an appropriate version of `mvn` is used (`maven-enforcer-plugin`).

Finally, improved meta-information included in the `pom.xml` files around
licenses and attribution, to be more homogenous & de-duplicated.

None of the changes beyond replacing `JSII-305` with the JetBrains
annotations is really material in terms of the contents of the packages,
as those are essentially meta-information changes only. They could be
extracted out if we feel like we should split changes up.