feat(jsii): files to exclude in tsconfig.json
This change lets package.json of a jsii package to specify additional
'exclude' paths. They are encoded into tsconfig.json.

Motivation
jsconfig is a common way to configure tools like eslint and jest. These
files have the 'js' extension but should not be compiled into the
package. This flag lets the jsii compiled packages to configure this.