fix(python): imports between subpackages are broken
Using absolute imports from a sub-package to another of the same root
package causes cycles in the module loading machinery, resulting in
obscure errors (typically, the loader complaining that the root module
name is not defined).

This changes the name generation mechanism so that relative imports are
used to refer to sub-packages within a same root, as per the best
practices. This allows the module loaded to flatten the loop and
successfully load everything (granted, as always, there is no
intractible loop between the sub-packages).

In addition, this change makes the `._jsii` submodule take care of
ensuring the current package's depencencies' assemblies are loaded up
before loading it's own module. This alleviated the need to
systematically import all dependencies from each sub-package, instead
only loading up the packages that are effectively used within a module.