fix(python): packages fail to load when relocated (#1518)
When a package is relocated from the root to within an arbitrary

subpackage, the loading infrastructure fails to locate the .tgz file
containing the JavaScript runtime for the library, because the original
library name is hard-coded in the package.

This removes the hardcoding and computes the modules' name from the
`__name__` obtained at runtime.

Additionally, moved the loading of the assembly within the `_jsii`
subpackage, which makes the overall generated code ever so slightly
cleaner.

Fixes #1501