chore(python): cosmetic improvements on generated code
Add an explicit `-> None` type annotation to the `__init__` function of struct constructors, to be consistent with the recommendations of [PEP-484](https://www.python.org/dev/peps/pep-0484/#the-meaning-of-annotations).

Consistently put a space before the comma in `typing.Mapping[]` arguments. Previously the space was missing in the case of Maps, but present in the case of JSON objects.

The increased consistency of generated code would facilitate future refactorings, as less special cases have to be accounted for.