fix(jsii): unable to use type from dependencies' submodules
When referring to a type defined in a submodule of a dependency (but
not using the "inline `namespace`" syntax style), the `jsii` compiler
failed to account for the namespace segment and generated an incorrect
`fqn`, leading to a type resolution error failing the compilation.

This adds a pass in the compiler to gather submodule layout from
dependencies and correctly register them in the compilation context, so
correct `fqn` are emitted.

Additionally, this highlighted an issue where namespaces defined using
the "inline `namespace`" syntax style would not be registered in the
`this._submodules` set, causing them to not always be properly
accounted for (depending on how the `TypeChecker` would represent
them).

New types have been introduced in `@scope/jsii-calc-lib` and `jsii-calc`
to verify that the compiler produces reasonable output. Those will be
leveraged in new compliance tests in a subsequent PR, as some
outstanding code-generation issues come in the way of these tests.