chore(dotnet): move all packages to single solution
The multiple solutions introduced for managing the DotNet runtime
packages result in unnecessary complexity in managing the mono-repo.
Consequently, they have been rolled into the `@jsii/dotnet-runtime`
package to make things easier going forward.

Since the projects are all on a single solution, it is also possible to
use a solution-wide declaration of dependency versions instead of
duplicating the values everywhere.