fix(java): compilation fails with "code too large"
In cases where a module includes a large amount of types, compilation of
Java packages could fail with the "code too large" error. This is due to
the presence of a single method in the `$Module` class that contains a
reference to all types in the package, resulting in a large amount of
generated bytecode (more than Java's maximum of 64KB).

In order to remediate, moved the FQN to Class mapping to a resource file
that gets loaded when the `$Module` class is initialized. The individual
`Class` entries are then resolved at the last minute using
`Class.forName` to ensure the reoslution happens *after* the full
mapping has been processed.