fix(jsii): dependency submodules are not tagged (#1663)
The Assembler used to ignore any export that was coined as an external
library import by the TypeScript resolver. However, when traversing
dependencies to collect their submodule structure, everything is an
external library import (which basically is a check that the import
path has a `/node_modules/` in it).

This commit changes the way external imports are skipped so that when
investigating submodules of a particular dependency, a different logic
is used to determine whether they are external (checks that the file
is under the same root; and that it is not under a `/node_modules/`
subdirectory of that root).

> **Note:**
> This is difficult to test for in the contect of a mono-repo, as it
> requires an external dependency (dependencies within the mono-repo are
> resolved to their "canonical" path, which does not include the
> `/node_modules/` path element, and thus is not reported as an external
> library import, apparently).