chore: make integration test use local runtimes (#1676)
Instead of performing local installs, set the environment variables
needed to direct the CDK build gear to use the locally built tools from
within their location in the mono-repository.

No doing this can cause parts of the tools to be installed from the NPM
package registry, which is not desirable.