chore(pacmak): fix local runtime detection
Local installs of the java runtime were not properly discovered because
the opportunistic resolution used the incorrect package name (it was
using the older `jsii-java-runtime` name instead of the new
`@jsii/java-runtime` name).

This commit fixes that problem, and also adds an `info` log entry that
indicates a locally built runtime is being used instead of the standard
one from the registry.