feat(jsii): use incremental builders
Using the `IncrementalBuilder` to compiler code instead of a regular
`Builder` allows saving time when certain files are aloready up-to-date.

The incremental support is only enabled when `--project-references` is
passed to the compiler, as it relies on project references and the
composite model to work it's magic (as far as `tsc` is concerned, the
`incremental` option implies the `composite` option).