fix(kernel): error raised during rename operation
When loading a new library into the jsii kernel, the provided `tarball`
was extracted to a temporary directory, then moved into it's final
location. On Windows, this operation could fail on an `EACCESS` or
`EPERM` error (often due to malware scanners accessing the file for
inspection on file systems which do not support renaming files that are
being accessed).

This changes how the `load` API works so taht the `tarball` is extracted
directly in it's final install location, so that no rename operation is
needed.