feat(pacmak/java): isolate maven repositories
When preparing a Java package using Maven, `jsii-pacmak` now configures
a dedicated `localRepository` for it's own use. This prevents artifacts
cached from previous builds as well as local build artifacts to possibly
interfere with the compilation process.

This has the side-effect to force a re-download of all dependencies on
each build (except those that are generated as part of the same build),
which results in a performance degradation. It should be possible to
make the local repository location user-configurable so this impact can
be mitigated on a case-by-case basis.