fix(dotnet): "Could not infer JSII type for .NET type 'AnonymousObject'"
When obtained through an untyped way (e.g: as part of an opaque object
for example), instances of `AnonymousObject` could not be cast back to
their Kernel form, because the converter lacked a code path to handle
this type. Instead, it insisted on trying to infer a "better" run-time
type for it.

This adds the necessary code path to handle this condition and properly
return the reverse conversion.

Fixes aws/aws-cdk#7977